+++ lib/spreedly/environment.rb
@@ -248,7 +248,7 @@ module Spreedly
       build_xml_request('payment_method') do |doc|
         add_to_doc(doc, options, :data, :retained, :email)
         doc.credit_card do
-          add_to_doc(doc, options, :number, :verification_value, :month, :first_name, :last_name,
+          add_to_doc(doc, options, :number, :verification_value, :month, :full_name, :first_name, :last_name,
                      :year, :address1, :address2, :city, :state, :zip, :country, :phone_number,
                      :company, :eligible_for_card_updater)
         end
@@ -257,7 +257,7 @@ module Spreedly
 
     def update_credit_card_body(options)
       build_xml_request('payment_method') do |doc|
-        add_to_doc(doc, options, :email, :month, :first_name, :last_name, :year,
+        add_to_doc(doc, options, :email, :month, :full_name, :first_name, :last_name, :year,
                    :address1, :address2, :city, :state, :zip, :country, :phone_number,
                    :eligible_for_card_updater)
       end
@@ -284,13 +284,23 @@ module Spreedly
 
     def add_extra_options_for_basic_ops(doc, options)
       add_gateway_specific_fields(doc, options)
-      add_to_doc(doc, options, :order_id, :description, :ip, :merchant_name_descriptor,
+      add_shipping_address_override(doc, options)
+      add_to_doc(doc, options, :order_id, :description, :ip, :email, :merchant_name_descriptor,
                                :merchant_location_descriptor)
     end
 
     def add_gateway_specific_fields(doc, options)
       return unless options[:gateway_specific_fields].kind_of?(Hash)
       doc << "<gateway_specific_fields>#{xml_for_hash(options[:gateway_specific_fields])}</gateway_specific_fields>"
+    end
+
+    def add_shipping_address_override(doc, options)
+      return unless options[:shipping_address].kind_of?(Hash)
+      doc.send(:shipping_address) do
+        options[:shipping_address].each do |k, v|
+          doc.send(k, v)
+        end
+      end
     end
 
     def xml_for_hash(hash)