+++ test/remote/remote_add_credit_card_test.rb
@@ -27,7 +27,7 @@ class RemoteAddCreditCardTest < Test::Unit::TestCase
   end
 
   def test_payment_required
-    assert_raise_with_message(Spreedly::PaymentRequiredError, "Your account has not been activated for real transactions. Please update your subscription settings.") do
+    assert_raise_with_message(Spreedly::PaymentRequiredError, "Your environment (R7lHscqcYkZeDGGbthKp6GKMu15) has not been activated for real transactions with real payment methods. If you're using a Test Gateway you can *ONLY* use Test payment methods - ( https://docs.spreedly.com/test-data). All other credit card numbers are considered real credit cards; real credit cards are not allowed when using a Test Gateway.") do
       @environment.add_credit_card(card_deets(number: '343'))
     end
   end
@@ -48,6 +48,14 @@ class RemoteAddCreditCardTest < Test::Unit::TestCase
     assert t.succeeded?
     assert t.retained
     assert_equal 'retained', t.payment_method.storage_state
+  end
+
+  def test_successfull_add_using_full_name
+    t = @environment.add_credit_card(number: '5555555555554444', month: 1, year: 2019, full_name: "Kvothe Jones")
+    assert t.succeeded?
+    assert_equal "Kvothe", t.payment_method.first_name
+    assert_equal "Jones", t.payment_method.last_name
+    assert_equal "Kvothe Jones", t.payment_method.full_name
   end